requirements.txt
openclean-core>=0.3.0
Shapely>=1.7.1
uszipcode>=0.2.4
